more entity objects

+++ src/main/java/app/persistence/entity/CombatStats.java
@@ -14,20 +14,20 @@ public class CombatStats {
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
 	
-	private long strength;
-	private long constitution;
-	private long endurance;
-	private long dexterity;
-	private long intellect;
-	private long wisdom;
-	private long charisma;
+	private Double strength;
+	private Double constitution;
+	private Double endurance;
+	private Double dexterity;
+	private Double intellect;
+	private Double wisdom;
+	private Double charisma;
 	
 	private List<Ailment> ailments;
 	
 	protected CombatStats() {
 	}
 	
-	public CombatStats(long strength, long constitution, long endurance, long dexterity, long intellect, long wisdom, long charisma) {
+	public CombatStats(Double strength, Double constitution, Double endurance, Double dexterity, Double intellect, Double wisdom, Double charisma) {
 		this.strength = strength;
 		this.constitution = constitution;
 		this.endurance = endurance;
